==> 04-Encoding_estandarizacion_sinout.ipynb <==
# ENCODING Y ESTANDARIZACIÓN
---

import pandas as pd

from sklearn.preprocessing import StandardScaler
import pickle
import warnings
warnings.filterwarnings('ignore')

df= pd.read_csv("datos/02-bikes_sinoutliers.csv", index_col= 0).reset_index(drop= True)
df.head()

### Probamos primero estas opciones
- **Year** : El encoding estaría bien hecho ya porque el año 1 hay más bicis alquiladas
- **Workingday** : El encoding también estaría bien
- **Weathersit** : El encoding sería 1 - 2, para el 2 le ponemos un 1 y para el 3 le ponemos un 0
- **Season** : Hay que hacerlo. Verano 2, primavera y otoño 1, winter 0
- **mnth** enero: 0, febrero, marzo, abril, noviembre y diciembre: 1, junio, julio, agosto, septiembre, octubre: 2
- **Weekday** : Hacemos un map diferente para registered y para casual

map_holiday = {0:7, 1:0}
df["holiday"] = df["holiday"].map(map_holiday)

map_weathersit = {3:0, 2:2, 1:4}
df["weathersit"] = df["weathersit"].map(map_weathersit)

map_season = {"winter":0, "autumn":1, "spring":1, "summer":2}
df["season"] = df["season"].map(map_season)

mapa_weekday = {1: 0, 2:1, 3: 1, 4: 2, 5: 2, 6:2, 7:1}
df["weekday"] = df["weekday"].map(mapa_weekday)

mapa_mnth = {1: 0, 2:0, 3: 1, 4: 1, 5: 2, 6:2, 7:2, 8:2, 9:2, 10:2, 11:1, 12:1}
df["mnth"] = df["mnth"].map(mapa_mnth)

map_workingday = {0:1, 1:2}
df["workingday"] = df["workingday"].map(map_workingday)

df.head()

df.to_csv("datos/06-bikes_encoding_sinest_sinout.csv")

df_est = df.copy()
df_est.head(1)

## ESTANDARIZACION

num = df_est[["atemp", "hum", "windspeed"]]
num.head(2)

st = StandardScaler()
x_tranformadas = pd.DataFrame(st.fit_transform(num))
x_tranformadas.columns = num.columns
df_est[x_tranformadas.columns] = x_tranformadas
df_est.head()

df_est.to_csv("datos/07-bikes_encoding_est_sinout.csv")

with open ("datos/estandarizar.pkl", "wb") as fp:
    pickle.dump(st, fp)
