==> 05-DecisionTree_RandomForest_sinout.ipynb <==
# DECISION TREE Y RANDOM FOREST
---

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import GradientBoostingRegressor
from sklearn import tree
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.model_selection import GridSearchCV

import pickle 
import warnings
warnings.filterwarnings('once')

df_est = pd.read_csv("datos/07-bikes_encoding_est_sinout.csv", index_col = 0)
df_est.head()

df_est.drop("temp", axis = 1, inplace= True)

df_est.season.unique()

df_est.isnull().sum()

# Creamos una clase para poder hacer el decision tree y el random forest más rápido
class Regres_lineal:
    def __init__(self, dataf, col_pred):
        
        self.dataf = dataf
        self.col_pred = col_pred

    def tree_param(self):
        X = self.dataf.drop(self.col_pred, axis = 1)
        y = self.dataf[self.col_pred]

        x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)

        arbol = DecisionTreeRegressor(random_state =0)
    
        arbol.fit(x_train, y_train)

        max_feat = np.sqrt(len(x_train.columns))
        max_dep = arbol.tree_.max_depth
        return(max_feat, max_dep)

    def metricas(self, y_te, y_tr, y_te_pred, y_tr_pred, tipo_modelo):
        resultados = {'MAE': [mean_absolute_error(y_te, y_te_pred), mean_absolute_error(y_tr, y_tr_pred)],
                    'MSE': [mean_squared_error(y_te, y_te_pred), mean_squared_error(y_tr, y_tr_pred)],
                    'RMSE': [np.sqrt(mean_squared_error(y_te, y_te_pred)), np.sqrt(mean_squared_error(y_tr, y_tr_pred))],
                    'R2':  [r2_score(y_te, y_te_pred), r2_score(y_tr, y_tr_pred)],
                    "set": ["test", "train"]}
        df = pd.DataFrame(resultados)
        df["modelo"] = tipo_modelo
        return df

    def grid_s_decision(self, dict, model):
        self.dict = dict
        self.model = model
        
        X = self.dataf.drop(self.col_pred, axis = 1)
        y = self.dataf[self.col_pred]
        x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)
        
        gs = GridSearchCV(
            estimator= GradientBoostingRegressor(), 
            param_grid= self.dict, 
            cv=10, 
            verbose=-1,
            return_train_score = True,
            scoring="neg_mean_squared_error")
        gs.fit(x_train, y_train)
        mejor_modelo = gs.best_estimator_
        y_pred_test_dt2 = mejor_modelo.predict(x_test)
        y_pred_train_dt2 = mejor_modelo.predict(x_train)
        dt_results = self.metricas(y_test, y_train, y_pred_test_dt2, y_pred_train_dt2, self.model)
        return mejor_modelo, dt_results

    def grid_s_forest(self, dict, model):
        self.dict = dict
        self.model = model
        
        X = self.dataf.drop(self.col_pred, axis = 1)
        y = self.dataf[self.col_pred]
        x_train, x_test, y_train, y_test = train_test_split(X, y, test_size = 0.2, random_state = 42)
        
        gs = GridSearchCV(
            estimator= RandomForestRegressor(), 
            param_grid= self.dict, 
            cv=10, 
            verbose=-1,
            return_train_score = True,
            scoring="neg_mean_squared_error")
        gs.fit(x_train, y_train)
        mejor_modelo = gs.best_estimator_
        y_pred_test_dt2 = mejor_modelo.predict(x_test)
        y_pred_train_dt2 = mejor_modelo.predict(x_train)
        dt_results = self.metricas(y_test, y_train, y_pred_test_dt2, y_pred_train_dt2, self.model)
        return mejor_modelo, dt_results


rg_est = Regres_lineal(df_est, "registered")

features_est, depth_est = rg_est.tree_param()

features_est #El número máx será 3

depth_est # La profundidad máxima será 22

param1 = {"max_depth": [5], 
        "max_features": [4],
        "min_samples_split": [50],
        "min_samples_leaf": [11]} 

best_model1, result1 = rg_est.grid_s_decision(param1, "GradientBoosting Estand I")

best_model1

result1

param2 = {"max_depth": [8], 
        "max_features": [4],
        "min_samples_split": [40],
        "min_samples_leaf": [5]} 

best_model2, result2 = rg_est.grid_s_decision(param2, "GradientBoosting Estand II")

best_model2

result2

param31 = {"max_depth": [4], 
        "max_features": [6],
        "min_samples_split": [50],
        "min_samples_leaf": [10]} 

best_model31, result31 = rg_est.grid_s_decision(param31, "GradientBoosting Estand III")

best_model31

with open ("datos/model_prediccion.pkl", "wb") as fp:
    pickle.dump(best_model31, fp)

result31

param32 = {"max_depth": [7], 
        "max_features": [5],
        "min_samples_split": [50],
        "min_samples_leaf": [10]} 

best_model32, result32 = rg_est.grid_s_decision(param32, "GradientBoosting Estand IV")

best_model32

result32

param3 = {"max_depth": [10], 
        "max_features": [5],
        "min_samples_split": [60],
        "min_samples_leaf": [15]} 

best_model3, result3 = rg_est.grid_s_forest(param3, "RandForest Estand I")

best_model3

result3

param4 = {"max_depth": [4], 
        "max_features": [3],
        "min_samples_split": [10],
        "min_samples_leaf": [50]} 

best_model4, result4 = rg_est.grid_s_forest(param4, "RandForest Estand II")

best_model4

result4

df_sinest = pd.read_csv("datos/06-bikes_encoding_sinest_sinout.csv", index_col = 0)
df_sinest.head()

df_sinest.drop("temp", axis= 1, inplace=True)

rg_sinest = Regres_lineal(df_sinest, "registered")

features_sinest, depth_sinest = rg_sinest.tree_param()

features_sinest #El número máx será 3

depth_sinest # La profundidad máxima será 22

best_model5, result5 = rg_sinest.grid_s_decision(param1, "GradientBoosting SinEstand I")

result5

best_model6, result6 = rg_sinest.grid_s_decision(param2, "GradientBoosting SinEstand II")

result6

best_model7, result7 = rg_sinest.grid_s_forest(param3, "RandForest SinEstand I")

result7

best_model8, result8 = rg_sinest.grid_s_forest(param4, "RandForest SinEstand II")

result8

param5 = {"max_depth": [4], 
        "max_features": [6],
        "min_samples_split": [50],
        "min_samples_leaf": [10]}

best_model9, result9 = rg_sinest.grid_s_forest(param5, "RandForest SinEstand III")

result9

best_model10, result10 = rg_est.grid_s_forest(param5, "RandForest Estand III")

result10

param6 = {"max_depth": [18,20], 
        "max_features": [3,4],
        "min_samples_split": [8, 15],
        "min_samples_leaf": [8, 15]}

best_model11, result11 = rg_est.grid_s_forest(param6, "RandForest Estand IV")

best_model11

result11

best_model12, result12 = rg_est.grid_s_decision(param5, "GradientBoosting Estand III")

result12

best_model13, result13 = rg_sinest.grid_s_decision(param5, "GradientBoosting SinEstand III")

result13

param7 = {"max_depth": [4], 
        "max_features": [3],
        "min_samples_split": [50],
        "min_samples_leaf": [10]}

best_model14, result14 = rg_est.grid_s_decision(param7, "GradientBoosting Estand IV")

best_model14

result14

param8 = {"max_depth": [4], 
        "max_features": [3],
        "min_samples_split": [40],
        "min_samples_leaf": [9]}

best_model15, result15 = rg_est.grid_s_decision(param8, "GradientBoosting Estand V")

best_model15

result15
