==> 06-Prediccion.ipynb <==
import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.tree import DecisionTreeRegressor
from sklearn.ensemble import RandomForestRegressor
from sklearn.ensemble import GradientBoostingRegressor
from sklearn import tree
from sklearn.metrics import r2_score, mean_squared_error, mean_absolute_error
from sklearn.model_selection import GridSearchCV
from sklearn.preprocessing import StandardScaler
from sklearn.preprocessing import OneHotEncoder
import pickle
import warnings
warnings.filterwarnings('once')

with open ("datos/estandarizar.pkl" , "rb") as fp:
    bike = pickle.load(fp)

- Si el año es 2018: 0
- Si el año es 2019: 1
- Si son vacaciones: 1
- Si no son vaciones: 0
- Si es día laborable: 1
- Si no es día laborable: 0
- Los días de la semana van del 1 al 7 empezando por lunes.
- Si weathersit = 1 -> sol
- Si weathersit = 2 -> nublado
- Si weathersit = 3 -> nieve, lluvia


usuario = {"season": "summer" , "yr": 1, "mnth": 8 ,"holiday": 0, "weekday": 5 ,"workingday":1, "weathersit": 1, "atemp": 40.5, "hum": 78, "windspeed": 10.9 }
df_usuario = pd.DataFrame(usuario, index = [0])
df_usuario

map_holiday = {0:7, 1:0}
df_usuario["holiday"] = df_usuario["holiday"].map(map_holiday)
map_weathersit = {3:0, 2:2, 1:4}
df_usuario["weathersit"] = df_usuario["weathersit"].map(map_weathersit)
map_season = {"winter":0, "autumn":1, "spring":1, "summer":2}
df_usuario["season"] = df_usuario["season"].map(map_season)
mapa_weekday = {1: 0, 2:1, 3: 1, 4: 2, 5: 2, 6:2, 7:1}
df_usuario["weekday"] = df_usuario["weekday"].map(mapa_weekday)
mapa_mnth = {1: 0, 2:0, 3: 1, 4: 1, 5: 2, 6:2, 7:2, 8:2, 9:2, 10:2, 11:1, 12:1}
df_usuario["mnth"] = df_usuario["mnth"].map(mapa_mnth)
map_workingday = {0:1, 1:2}
df_usuario["workingday"] = df_usuario["workingday"].map(map_workingday)

numericas = df_usuario[["atemp", "hum", "windspeed"]]
numericas

df_bike = pd.DataFrame(bike.transform(numericas))


df_usuario[["atemp", "hum", "windspeed"]] = df_bike

with open ("datos/model_prediccion.pkl" , "rb") as fs:
    modelo = pickle.load(fs)

modelo.predict(df_usuario)
